Linux work...

+++ include/lgfx/Linxc.h
@@ -1,6 +1,8 @@
 #ifndef LINXC_H
 #define LINXC_H
 
+#include <stddef.h>
+
 #ifndef NULL
 #define NULL 0
 #endif
@@ -38,7 +40,7 @@ typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned int u32;
 typedef unsigned long long u64;
-typedef unsigned long long usize;
+typedef size_t usize;
 
 typedef const char *text;
 typedef const char *u8text;
